Update
ML models have been trained and have been compared on basis of their accuracy, precision and recall.

NB: Code in python script is a little different from the code in the R Markdown document. Final output is shown in the html document. The code in Python script was rough work.